requirements.txt
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.1
Werkzeug==3.0.0
bcrypt==4.1.0
alembic==1.16.4
PyMySQL==1.1.1
boto3==1.40.16